TRSN-10: Applying mobile styles to the QuizResult page
- Added mobile detection logic to QuizResults component
- Implemented responsive styling for mobile devices (screen width < 1024px)
- Added window resize event listener for dynamic mobile state updates
- Simplified component interface by removing unnecessary isMobile prop
- Made QuizResults self-contained with internal mobile state management
- Ensured consistent mobile styling across quiz and results pages
- Added proper cleanup for resize event listeners to prevent memory leaks

Fixes #10
