+++ quiz-viewer/package.json
@@ -23,6 +23,7 @@
     "react-dnd": "^16.0.1",
     "react-dnd-html5-backend": "^16.0.1",
     "react-dom": "^19.1.1",
+    "react-router-dom": "^7.8.2",
     "tailwind-merge": "^3.3.1"
   },
   "devDependencies": {
